removed useless files, to avoid having more files than needed
